價格類型改為支援浮點數 %.2f

+++ handlers/superchat_msg.go
@@ -63,7 +63,7 @@ func superChatMsg(data *blive.LiveData) error {
 		TargetUid: targetUid,
 		Command:   blive.SuperChatMessage,
 		Display:   display,
-		Price:     price,
+		Price:     float64(price),
 	}
 
 	result := db.Database.Create(behaviour)
